remove unnessary casts
git-svn-id: https://ncisvn.nci.nih.gov/svn/coppa/trunk/code@5442 553898ca-32d5-487a-af8b-6cf4ca579ab4

+++ coppa-commons/src/main/java/gov/nih/nci/coppa/services/grid/dto/transform/iso/INTTransformer.java
@@ -44,7 +44,7 @@ public final class INTTransformer extends QTYTransformer<INT, Int> implements Tr
         if (input == null) {
             return null;
         }
-        INT x = (INT) transformBaseXml(input);
+        INT x = transformBaseXml(input);
         Integer v = input.getValue();
         if (v != null) {
             x.setValue(v);
@@ -62,7 +62,7 @@ public final class INTTransformer extends QTYTransformer<INT, Int> implements Tr
         if (input == null) {
             return null;
         }
-        Int d = (Int) transformBaseDto(input);
+        Int d = transformBaseDto(input);
         Integer v = input.getValue();
         if (v != null) {
             d.setValue(v);
